fix: bug with negative length
len < 0 is always false,  because len is a uint and uint can never be negative.

The original C version casts to int here, so that the uint is converted to 2's compliment negative numbers.

+++ kcp2k/Assets/kcp2k/kcp/Kcp.cs
@@ -512,7 +512,7 @@ namespace kcp2k
 
                 size -= OVERHEAD;
 
-                if (size < len || len < 0) return -2;
+                if (size < (int)len || (int)len < 0) return -2;
 
                 if (cmd != CMD_PUSH && cmd != CMD_ACK &&
                     cmd != CMD_WASK && cmd != CMD_WINS)
